% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sp.R
\name{as.data.frame.ejAttribute}
\alias{as.data.frame.ejAttribute}
\title{Convert an ej attribute to a dataframe}
\usage{
as.data.frame.ejAttribute(x)
}
\arguments{
\item{x}{An ejAttribute}
}
\description{
Convert an ej attribute to a dataframe
}

