% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.frame.R
\name{as.ejObject.data.frame}
\alias{as.ejObject.data.frame}
\title{Convert a dataframe to an ejObject}
\usage{
\method{as.ejObject}{data.frame}(x, recordID = NA, recordAttributes,
  eventDefinitions, metadata = list(), ...)
}
\arguments{
\item{x}{The dataframe to convert}

\item{recordID}{an ID for the record, if NA one is created}

\item{recordAttributes}{A character vector containing the names of the
columns in the dataframe that are attributes of the record}

\item{eventDefinitions}{A list of event definitions}

\item{metadata}{A list of metadata ejAttribute objects describing the dataset}

\item{...}{other parameters (to maintain consistency with the generic)}
}
\description{
Convert a dataframe to an ejObject
}
\note{
We assume one row per record.
}
\examples{
library(HistData)
 data(Snow.deaths)
 #It would be good to have some dates, pumps, some genders
 simulated <- Snow.deaths
 simulated$gender <- c("male","female")[(runif(nrow(simulated))>0.5) +1]
 simulated$date <- as.POSIXct("1854-04-05") + rnorm(nrow(simulated), 10) * 86400
 simulated$pump <- ceiling(runif(nrow(simulated)) * 5)
 as.ejObject(simulated, recordAttributes = c("gender"),
		eventDefinitions = list(defineEjEvent(dateStart="date", dateEnd="date", name=NA, location=list(x="x", y="y", proj4string=""), attributes="pump")),
		metadata=list())
}

