% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epiJSON.R
\name{as.ejObject.default}
\alias{as.ejObject.default}
\title{By default we don't know how to convert objects}
\usage{
\method{as.ejObject}{default}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{other parameters passed to the call}
}
\description{
By default we don't know how to convert objects
}

