% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sp.R
\name{as.SpatialPointsDataFrame.ejEvent}
\alias{as.SpatialPointsDataFrame.ejEvent}
\title{convert an ejEvent to a SpatialPointsDataFrame}
\usage{
as.SpatialPointsDataFrame.ejEvent(x)
}
\arguments{
\item{x}{An ejEvent object}
}
\description{
convert an ejEvent to a SpatialPointsDataFrame
}

