% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sp.R
\name{as.SpatialPointsDataFrame.ejObject}
\alias{as.SpatialPointsDataFrame.ejObject}
\title{Create a SpatialPointsDataFrame from an ejObject}
\usage{
as.SpatialPointsDataFrame.ejObject(x)
}
\arguments{
\item{x}{An ejObject}
}
\description{
Create a SpatialPointsDataFrame from an ejObject
}

