% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sp.R
\name{as.SpatialPointsDataFrame.ejRecord}
\alias{as.SpatialPointsDataFrame.ejRecord}
\title{convert an ejRecord to a SpatialPointsDataFrame}
\usage{
as.SpatialPointsDataFrame.ejRecord(x)
}
\arguments{
\item{x}{an ejRecord object}
}
\description{
convert an ejRecord to a SpatialPointsDataFrame
}

