% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epijsonObjectVis.R
\name{box2}
\alias{box2}
\title{generic labelled box function}
\usage{
box2(xmin = 0.01, xmax = 0.99, ymin = 0.01, ymax = 0.99,
  label = "box2", colour = "gray60", fill = "white", gg = NULL,
  addSheets = 0, size = 4, print = FALSE)
}
\arguments{
\item{xmin}{0 to 1}

\item{xmax}{0 to 1}

\item{ymin}{0 to 1}

\item{ymax}{0 to 1}

\item{label}{to print at top of box}

\item{colour}{border of box}

\item{fill}{fill colour for box NA=none}

\item{gg}{ggplot object, if passed box is added if NULL new ggplot object created}

\item{addSheets}{how many sheets to add behind box to indicate array}

\item{size}{font size for label}

\item{print}{TRUE/FALSE whether to print the ggplot object}
}
\value{
a ggplot object
}
\description{
generic labelled box function
}

