% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constructors.R
\name{createAttribute}
\alias{createAttribute}
\title{Create an attribute
This package outlines the aspects of the data for EpiJSON}
\usage{
createAttribute(name, type, value)
}
\arguments{
\item{name}{This is the name of the attribute, usually a column name}

\item{type}{This is the type of information, either string, float, integer, boolean or date}

\item{value}{This is the value that the data takes for the given attribute}
}
\value{
an ejAttribute object
}
\description{
This function defines attributes
output \code{ejAttribute}
}
\examples{
dF<- data.frame(id=c("A","B","3D"),
                name=c("tom","andy","ellie"),
                dob=c("1984-03-14","1985-11-13","1987-06-16"),
                gender=c("male","male","female"),
                rec1contact=c(2,1,5),
                rec1dateStart=c("2014-12-28","2014-12-29","2015-01-03"),
                rec1dateEnd=c("2014-12-30","2015-01-04","2015-01-07"),
                rec1risk=c("high","high","low"),
                rec1temp=c(39,41,41),
                rec2contact=c(4,1,1),
                rec2date=c("2015-01-02","2015-01-12","2015-01-09"),
                rec2risk=c("high","low","high"),stringsAsFactors=FALSE)

createAttribute(name="name",type="int",value=dF$name[1])
}

