% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/constructors.R
\name{createEvent}
\alias{createEvent}
\title{Create an event}
\usage{
createEvent(id = NA, name, dateStart, dateEnd, location, attributes)
}
\arguments{
\item{id}{This is the identifier for the event}

\item{name}{This is the name of the event, usually a column name}

\item{dateStart}{This is the (inclusive) date on which this event
started (or occured).}

\item{dateEnd}{This is the (exclusive) date on which this event finished. A
point event in time would have a dateStart and a dateEnd that are identical.}

\item{location}{This is the location at which this event happened}

\item{attributes}{This is a concatenated list of attributes associated with this event}
}
\value{
an ejEvent object
}
\description{
This function defines events
output \code{ejEvent}
}
\examples{
dF<- data.frame(id=c("A","B","3D"),
                name=c("tom","andy","ellie"),
                dob=c("1984-03-14","1985-11-13","1987-06-16"),
                gender=c("male","male","female"),
                rec1contact=c(2,1,5),
                rec1dateStart=c(as.POSIXct(tz="GMT","2014-12-28"),as.POSIXct(tz="GMT","2014-12-29"),as.POSIXct(tz="GMT","2015-01-03")),
                rec1dateEnd=c(as.POSIXct(tz="GMT","2014-12-30"),as.POSIXct(tz="GMT","2015-01-04"),as.POSIXct(tz="GMT","2015-01-09")),
                rec1risk=c("high","high","low"),
                rec1temp=c(39,41,41),
                rec2contact=c(4,1,1),
                rec2date=c("2015-01-02","2015-01-12","2015-01-09"),
                rec2risk=c("high","low","high"),stringsAsFactors=FALSE)

event1<-createEvent(id=NA,
                      name="rec1contact",
                      dateStart=dF$rec1dateStart[1],
                      dateEnd=dF$rec1dateEnd[1],
                      location="",
                      attributes=list(createAttribute(name="rec1risk",type="str",value=dF$rec1risk[1]),
                             createAttribute(name="rec1temp",type="int",value=dF$rec1temp[1])))
}

