% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{dataFrameToAttributes}
\alias{dataFrameToAttributes}
\title{The general functions that are used in multiple sections of the epiJSON package}
\usage{
dataFrameToAttributes(x)
}
\arguments{
\item{x}{A dataframe}
}
\value{
result A list of attributes that comprise the dataframe
Convert a dataframe to attributes
}
\description{
The general functions that are used in multiple sections of the epiJSON package
}
\examples{
#commenting out example
#i think example fails check because the function isn't exported
}

