% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epiJSON.R
\docType{package}
\name{EpiJSON}
\alias{EpiJSON}
\alias{EpiJSON-package}
\title{A data format for epidemiological data}
\description{
a JSON format for epidemiological data with functions to convert to and from other useful formats
}

