% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epiJSON2r.r
\name{epiJSON2r}
\alias{epiJSON2r}
\title{converting epiJSON into a list of lists}
\usage{
epiJSON2r(file)
}
\arguments{
\item{file}{an epiJSON filename or string to convert to R}
}
\value{
a list of lists of the epijson content
}
\description{
takes an epiJSON string or file and converts to a list of lists
later this needs to be made more formal
}
\examples{
listJSON <- epiJSON2r( system.file("extdata//example.JSON", package="EpiJSON"))
str(listJSON)
#from within the package would do this
#listJSON <- epiJSON2r("extdata//example.JSON")
}

