% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epijsonObjectVis.R
\name{epijsonObjectVis}
\alias{epijsonObjectVis}
\title{to view the structure of epijson objects and/or schema
testing ability/utility
structure is largely hardcoded for now, could try to automate later}
\usage{
epijsonObjectVis()
}
\value{
a ggplot object
}
\description{
to view the structure of epijson objects and/or schema
testing ability/utility
structure is largely hardcoded for now, could try to automate later
}

