% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epijsonObjectVis.R
\name{epijsonObjectVis2}
\alias{epijsonObjectVis2}
\title{2 view the structure of epijson objects and/or schema
starting to automate, making box positions relative}
\usage{
epijsonObjectVis2()
}
\value{
a ggplot object
}
\description{
2 view the structure of epijson objects and/or schema
starting to automate, making box positions relative
}

