% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/epijsonObjectVis.R
\name{epijsonObjectVis3}
\alias{epijsonObjectVis3}
\title{3 view the structure of epijson objects and/or schema
allowing multiple attribute boxes with different labels}
\usage{
epijsonObjectVis3(attribMeta = "Attribute [name, type, value, units]",
  attribRecord = "Attribute [name, type, value, units]",
  attribEvent = "Attribute [name, type, value, units]")
}
\arguments{
\item{attribMeta}{label(s) for Metadata attributes}

\item{attribRecord}{label(s) for Record attributes}

\item{attribEvent}{label(s) for Event attributes}
}
\value{
a ggplot object
}
\description{
3 view the structure of epijson objects and/or schema
allowing multiple attribute boxes with different labels
}
\examples{
#this gives the base schema
epijsonObjectVis3()
#this gives a diagram with named attributes
epijsonObjectVis3( attribMeta = c("attribute: disease","attribute: data provider"),
                   attribRecord = c("attribute: gender","attribute: date of birth"),
                   attribEvent = c("attribute: recorder","attribute: test used") )
epijsonObjectVis3( attribMeta = c("a"),
                   attribRecord = c("b","c"),
                   attribEvent = c("d","e","f") )
}

