% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.frame.R
\name{findOrAdd}
\alias{findOrAdd}
\title{helper func to find a column name in a dataframe or add a new one}
\usage{
findOrAdd(dF, name, rowNum, value)
}
\arguments{
\item{dF}{a dataframe}

\item{name}{a column name to search for in the dataframe}

\item{rowNum}{which row to add the value to}

\item{value}{the value to put in the dataframe at the chosen row and column}
}
\description{
Checks for 'name' as a column name in the dataframe, if it is found the value is put there.
If the name is not found a new column is created and the value is put in the new column.
}

