% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.frame.R
\name{findOrAddAttributes}
\alias{findOrAddAttributes}
\title{helper func to go through all attributes and add them to a dataframe}
\usage{
findOrAddAttributes(dF, atts)
}
\arguments{
\item{dF}{a dataframe}

\item{atts}{an ejAttributes object}
}
\description{
DOESN'T WORK at the moment, so not called
adds 'new' attributes to a new column, existing attributes to existing column
}

