% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{notNA}
\alias{notNA}
\title{Return a value only if another is not NA}
\usage{
notNA(x, trueValue)
}
\arguments{
\item{x}{The value to test for NA}

\item{trueValue}{The value to return if x is not NA}
}
\value{
NA if x is NA or trueValue if x is not NA
}
\description{
Return a value only if another is not NA
}

