% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.R
\name{print.ejEvent}
\alias{print.ejEvent}
\title{print an ejEvent object}
\usage{
\method{print}{ejEvent}(x, ...)
}
\arguments{
\item{x}{An ejEvent object}

\item{...}{Other arguments to print (not used)}
}
\description{
print an ejEvent object
}
\examples{
dF<- data.frame(id=c("A","B","3D"),
                name=c("tom","andy","ellie"),
                dob=c("1984-03-14","1985-11-13","1987-06-16"),
                gender=c("male","male","female"),
                rec1contact=c(2,1,5),
                rec1dateStart=c("2014-12-28","2014-12-29","2015-01-03"),
                rec1dateEnd=c("2014-12-30","2015-01-04","2015-01-07"),
                rec1risk=c("high","high","low"),
                rec1temp=c(39,41,41),
                rec2contact=c(4,1,1),
                rec2date=c("2015-01-02","2015-01-12","2015-01-09"),
                rec2risk=c("high","low","high"),stringsAsFactors=FALSE)

x <- createEvent(id=NA,
             name="rec1contact",
             dateStart=as.POSIXct(dF$rec1dateStart[1]),
             dateEnd=as.POSIXct(dF$rec1dateEnd[1]),
             location=NA,
             attributes=list(createAttribute(name="rec1risk",type="str",value=c(as.character(dF$rec1risk[1]))),
                             createAttribute(name="rec1temp",type="int",value=c(as.character(dF$rec1temp[1])))))

print.ejEvent(x)
}

