% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.R
\name{print.ejMetadata}
\alias{print.ejMetadata}
\title{print an ejMetadata object}
\usage{
\method{print}{ejMetadata}(x, ...)
}
\arguments{
\item{x}{An ejMetadata object}

\item{...}{Other arguments to print (not used)}
}
\description{
print an ejMetadata object
}
\examples{
dF<- data.frame(id=c("A","B","3D"),
                name=c("tom","andy","ellie"),
                dob=c("1984-03-14","1985-11-13","1987-06-16"),
                gender=c("male","male","female"),
                rec1contact=c(2,1,5),
                rec1dateStart=c("2014-12-28","2014-12-29","2015-01-03"),
                rec1dateEnd=c("2014-12-30","2015-01-04","2015-01-07"),
                rec1risk=c("high","high","low"),
                rec1temp=c(39,41,41),
                rec2contact=c(4,1,1),
                rec2dateStart=c("2015-01-02","2015-01-12","2015-01-09"),
                rec2risk=c("high","low","high"),stringsAsFactors=FALSE)


x <- createMetadata (attributes=list(createAttribute(name="name",type="str",value=dF$name),
                                           createAttribute(name="dob",type="date",value=dF$dob),
                                           createAttribute(name="gender",type="str",value=dF$gender),
                                           createAttribute(name="rec1risk",type="str",value=dF$rec1risk),
                                           createAttribute(name="rec1temp",type="int",value=dF$rec1temp),
                                           createAttribute(name="rec2risk",type="str",value=dF$rec2risk)))

print(x)
}

