% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.R
\name{print.ejRecord}
\alias{print.ejRecord}
\title{print an ejRecord object}
\usage{
\method{print}{ejRecord}(x, ...)
}
\arguments{
\item{x}{An ejRecord object}

\item{...}{Other arguments to print (not used)}
}
\description{
print an ejRecord object
}
\examples{
dF<- data.frame(id=c("A","B","3D"),
                name=c("tom","andy","ellie"),
                dob=c("1984-03-14","1985-11-13","1987-06-16"),
                gender=c("male","male","female"),
                rec1contact=c(2,1,5),
                rec1dateStart=c("2014-12-28","2014-12-29","2015-01-03"),
                rec1dateEnd=c("2014-12-30","2015-01-04","2015-01-07"),
                rec1risk=c("high","high","low"),
                rec1temp=c(39,41,41),
                rec2contact=c(4,1,1),
                rec2date=c("2015-01-02","2015-01-12","2015-01-09"),
                rec2risk=c("high","low","high"),stringsAsFactors=FALSE)


x <- createRecord(id=dF$id[1],
             attributes=list(createAttribute(name="name",type="str",value=dF$name[1]),
                             createAttribute(name="dob",type="date",value=dF$dob[1]),
                             createAttribute(name="gender",type="str",value=dF$gender[1])),
             events=list(createEvent(id=NA,
                                       name="rec1contact",
                                       dateStart=as.POSIXct(dF$rec1dateStart[1]),
                                       dateEnd=as.POSIXct(dF$rec1dateEnd[1]),
                                       location="",
                                       attributes=list(createAttribute(name="rec1risk",type="str",value=dF$rec1risk[1]),
                                                       createAttribute(name="rec1temp",type="int",value=dF$rec1temp[1])
                                                                 )),
                         createEvent(id=NA,
                                      name="rec2contact",
                                      dateStart=dF$rec2dateStart[1],
                                      dateEnd=dF$rec2dateStart[1],
                                      location="",
                                             attributes=list(createAttribute(name="rec2risk",type="str",value=dF$rec2risk[1]))
                                       )
                             )
                     )


print(x)
}

