% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/obkData.R
\name{processRecord}
\alias{processRecord}
\title{This function takes a single record from the obkClass data and
converts to the epiJSON format}
\usage{
processRecord(x)
}
\arguments{
\item{x}{An record from the obkData}
}
\value{
an ejRecord
}
\description{
This function takes a single record from the obkClass data and
converts to the epiJSON format
}
\examples{
\dontrun{
#because this function is not exported this example won't work outside the package
require('OutbreakTools')
data(ToyOutbreak)
x <- subset(ToyOutbreak,1)
processRecord(x)
}
}

