% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/obkData.R
\name{processRecordFrame}
\alias{processRecordFrame}
\title{This function processes events from the obkClass data and
converts to the epiJSON format}
\usage{
processRecordFrame(x, recordFrameName)
}
\arguments{
\item{x}{An record from the obkData}

\item{recordFrameName}{The event of interest}
}
\value{
an ejEvent
Process an record event frame
}
\description{
This function processes events from the obkClass data and
converts to the epiJSON format
}
\examples{
##from utils.R run the dataFrameToAttributes function

##install OutbreakTools package from CRAN repository
install.packages(OutbreakTools)
require('OutbreakTools')

##An example dataset is available:
   data(ToyOutbreak)

   x=subset(ToyOutbreak,2)@events[[1]]

   processeventFrame(x,"Fever")
}

