% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r2epiJSON.r
\name{r2epiJSON}
\alias{r2epiJSON}
\title{Produce a JSON string from an EpiJSON object}
\usage{
r2epiJSON(x)
}
\arguments{
\item{x}{An EpiJSON object}
}
\value{
A character string in EpiJSON format.
}
\description{
Produce a JSON representation of an EpiJSON object. This should conform to
the EpiJSON standard.
}

