% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.frame.R
\name{as.data.frame.ejObject}
\alias{as.data.frame.ejObject}
\title{convert an \code{\link{ejObject}} to a dataframe with one row per record}
\usage{
as.data.frame.ejObject(ejOb)
}
\arguments{
\item{ejOb}{an \code{\link{ejObject}}}
}
\value{
dataframe
}
\description{
convert an \code{\link{ejObject}} to a dataframe with one row per record
}
\examples{
att <- createAttribute(name="name",type="int",value=7)
atts <- list(att)
event <- createEvent(id=NA, name="event1", dateStart=NA, dateEnd=NA, location=list(x="x", y="y", proj4string=""), attributes=atts )
#a single record
#records <- list(createRecord(id="bob", attributes=atts, events=list(event)))
ind1 <- createRecord(id="bob", attributes=atts, events=list(event))
ind2 <- createRecord(id="pat", attributes=atts, events=list(event))
records <- list(ind1,ind2)
ejOb <- createEJObject(metadata=atts, records=records)

#ejOb <- createEJObject(metadata=NULL, records=NULL)
as.data.frame(ejOb)

 #see eg from as.ejObject.data.frame for more complex example
 library(HistData)
 data(Snow.deaths)
 #It would be good to have some dates, pumps, some genders
 simulated <- Snow.deaths
 simulated$gender <- c("male","female")[(runif(nrow(simulated))>0.5) +1]
 simulated$dateStart <- as.POSIXct("1854-04-05") + rnorm(nrow(simulated), 10) * 86400
 simulated$dateEnd <- as.POSIXct("1854-04-05") + rnorm(nrow(simulated), 10) * 86400
 simulated$pump <- ceiling(runif(nrow(simulated)) * 5)
 ejOb2 <- as.ejObject(simulated, recordAttributes = c("gender"),
    eventDefinitions = list(defineEjEvent(dateStart="dateStart", dateEnd="dateEnd", name=NA, location=list(x="x", y="y", proj4string=""), attributes="pump")),
		metadata=list())
 as.data.frame(ejOb2)
}

