% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/obkData.R
\name{processRecord}
\alias{processRecord}
\title{This function takes a single record from the obkClass data and
converts to the epiJSON format}
\usage{
processRecord(x)
}
\arguments{
\item{x}{An record from the obkData}
}
\value{
an ejRecord
}
\description{
This function takes a single record from the obkClass data and
converts to the epiJSON format
}
\examples{
##from utils.R run the dataFrameToAttributes function

##install OutbreakTools package from CRAN repository
install.packages(OutbreakTools)
require('OutbreakTools')

##An example dataset is available:
     data(ToyOutbreak)

     x <- subset(ToyOutbreak,1)

     processrecord(x)
}

