% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/obkData.R
\name{processRecordFrame}
\alias{processRecordFrame}
\title{This function processes events from the obkClass data and
converts to the epiJSON format}
\usage{
processRecordFrame(x, recordFrameName)
}
\arguments{
\item{x}{An record from the obkData}

\item{recordFrameName}{The event of interest}
}
\value{
an ejEvent
}
\description{
This function processes events from the obkClass data and
converts to the epiJSON format
}
\examples{
\dontrun{
#because this function is not exported this example won't work outside the package
require('OutbreakTools')

data(ToyOutbreak)
x=subset(ToyOutbreak,2)@records[[1]]
processeventFrame(x,"Fever")
}
}

